requirements.txt
openai==1.6.0
python-dotenv==1.0.0
fastapi==0.110.0
uvicorn==0.29.0
